requirements.txt
aiogram==3.5.0
apscheduler==3.10.4
sqlalchemy==2.0.28
psycopg2-binary==2.9.9
python-dotenv==1.0.0
aiohttp==3.9.3
aiohttp-socks==0.8.4
beautifulsoup4==4.12.3
lxml==5.2.1
pybase64==1.3.2
